feat: Implement create order functionality and update customer home screen
This commit introduces the following changes:

-   **Create Order Functionality:**
    -   Implemented `CreateOrderScreen` composable to allow customers to create new orders with special instructions and items.
    -   Added `createOrder` function in `CustomerViewModel` to handle the order creation logic.
    -   Added `createOrder` function in `CustomerRepository` and `CustomerRepositoryImpl` to interact with the API service.
    -   Added `createOrder` API endpoint in `ApiService`.
    -   Added `CreateOrderRequest` and `CreateOrderResponse` data classes to handle the request and response for order creation.
    -   Added fake implementation for `createOrder` in `FakeCustomerRepositoryImpl` for testing purposes.
-   -  Update `CustomerAccountScreen` to call `CreateOrderScreen`.
+
+    **Customer Home Screen Update:**
+    -   Implemented shop listing in `CustomerHomeContent` by fetching data from the API.
+    -   Displayed shop information in a grid layout using `LazyVerticalGrid`.
+    -   Handled different states (Idle, Loading, Success, Error) for shop search using a `when` statement.
+    -   Displayed a `CircularProgressIndicator` while loading shop data.
+
+    **Navigation:**
+    -   Added `CustomerOrder` route to `LaundryAppScreen`.
+
+    **Other Changes:**
+    -   Updated imports and dependencies in various files to support the new functionality.
+    -   Removed unnecessary dependencies.
+    -   Updated the login state for customer screens.

+++ app/src/main/java/com/example/manage/laundry/data/model/response/CreateOrderResponse.kt
@@ -5,7 +5,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 data class CreateOrderResponse(
     val orderId: Int,
-    val totalPrice: Int,
+    val totalPrice: Double,
     val status: String,
     val createdAt: String
 )
