+++ app/src/main/java/com/example/manage/laundry/data/model/response/ShopOrderResponse.kt
@@ -7,7 +7,7 @@ import kotlinx.serialization.Serializable
 data class ShopOrderResponse(
     val orderId: Int,
     val customerName: String,
-    val totalPrice: Int,
+    val totalPrice: Double,
     val status: Order.Status,
     val createdAt: String
 )
